refactor: extract html block node
