+++ src/nodes/paragraph.ts
@@ -2,77 +2,8 @@ import type { RefDef, TsmarkNode } from '../types.d.ts';
 import { inlineToHTML } from './inline.ts';
 import { parseSetextHeading } from './heading.ts';
 import { parseThematicBreak } from './thematic_break.ts';
+import { htmlBlockStartRegex, htmlBlockTags } from './html.ts';
 import { indentWidth, LAZY, stripColumns, stripLazy } from '../utils.ts';
-
-const htmlBlockStartRegex = /^ {0,3}<\/?([A-Za-z][A-Za-z0-9-]*)(?=[\s/>]|$)/;
-const htmlBlockTags = new Set([
-  'address',
-  'article',
-  'aside',
-  'base',
-  'basefont',
-  'blockquote',
-  'body',
-  'caption',
-  'center',
-  'col',
-  'colgroup',
-  'dd',
-  'details',
-  'dialog',
-  'dir',
-  'div',
-  'dl',
-  'dt',
-  'fieldset',
-  'figcaption',
-  'figure',
-  'footer',
-  'form',
-  'frame',
-  'frameset',
-  'h1',
-  'h2',
-  'h3',
-  'h4',
-  'h5',
-  'h6',
-  'head',
-  'header',
-  'hr',
-  'html',
-  'iframe',
-  'legend',
-  'li',
-  'link',
-  'main',
-  'menu',
-  'menuitem',
-  'nav',
-  'noframes',
-  'ol',
-  'optgroup',
-  'option',
-  'p',
-  'param',
-  'pre',
-  'script',
-  'search',
-  'section',
-  'summary',
-  'style',
-  'table',
-  'tbody',
-  'td',
-  'tfoot',
-  'th',
-  'thead',
-  'title',
-  'tr',
-  'track',
-  'ul',
-  'textarea',
-]);
 
 export function parseParagraph(
   lines: string[],
